+++ packages/adapters/src/model/index.ts
@@ -1,3 +1,4 @@
 export * from "./client.js";
 export * from "./driver.js";
 export * from "./runtime.js";
+export * from "./transcript.js";